Extract CAN peripheral handler into separate module can_handler. This enables a separation of low-level device control from high-level business logic.
