+++ include/can.h
@@ -40,6 +40,16 @@ extern "C" {
 #define X8H7_CAN_HEADER_SIZE        5
 #define X8H7_CAN_FRAME_MAX_DATA_LEN	8
 
+/* Special address description flags for the CAN_ID */
+#define CAN_EFF_FLAG 0x80000000U /* EFF/SFF is set in the MSB */
+#define CAN_RTR_FLAG 0x40000000U /* remote transmission request */
+#define CAN_ERR_FLAG 0x20000000U /* error message frame */
+
+/* Valid bits in CAN ID for frame formats */
+#define CAN_SFF_MASK 0x000007FFU /* standard frame format (SFF) */
+#define CAN_EFF_MASK 0x1FFFFFFFU /* extended frame format (EFF) */
+#define CAN_ERR_MASK 0x1FFFFFFFU /* omit EFF, RTR, ERR flags */
+
 /**************************************************************************************
  * TYPEDEF
  **************************************************************************************/
